Renaming mixer hardware-layer files to be a bit more intuitive
